+++ ccg_frontend/ccg_mobile/utils/ParsingDuration.js
